Работа с ошибками Joi_9

+++ controllers/cards.js
@@ -7,6 +7,7 @@ import {
   NotFoundError,
   ConflictError,
   BadRequestError,
+  ForbiddenError,
 } from '../errors/index.js';
 
 const notFoundError = new NotFoundError('Запрашиваемый пользователь не найден');
@@ -14,6 +15,7 @@ const buildErrorServer = (message) => new ServerError(message);
 const buildErrorBadRequest = (message) => new BadRequestError(`Некорректные данные для пользователя. ${message}`);
 const errorNotUnique = new ConflictError('Пользователь с такой почтой уже существует');
 const UniqueErrorCode = 11000;
+const forbiddenError = new ForbiddenError('Эту карточку нельзя удалить, карточка другого пользователя');
 
 export function getAllCards(req, res) {
   Card.find({})
@@ -41,27 +43,67 @@ export function createCard(req, res) {
     });// данные не записались, вернём ошибку
 }
 
-export function deleteCard(req, res) {
-  Card.findOneAndRemove({ _id: req.params.id, owner: req.user._id })
-    .then((card) => ((!card) ? responseNotFound(res, 'Карточка не найдена, или это не карточка другого пользователя') : res.send(card)))
-    .catch((err) => {
-      if (err.name === 'ValidationError' || err.name === 'CastError') {
-        responseBadRequestError(res);
-      } else {
-        responseServerError(res);
-      }
-    });// данные не записались, вернём ошибку
+// export const deleteCard = async (req, res) => {
+//   try {
+//     const card = await Card.findById(req.params.id);
+//     if (!card) {
+//       responseNotFound(res, 'Карточка не найдена');
+//     } else if (card.owner !== req.user._id) {
+//       forbiddenError(res);
+//     } else {
+//       res.send(await Card.findOneAndRemove({ _id: req.params.id, owner: req.user._id }));
+//     }
+//   }
+//   catch (err) {
+//     if (err.name === 'ValidationError' || err.name === 'CastError') {
+//       responseBadRequestError(res);
+//     } else {
+//       responseServerError(res);
+//     }
+//   }
+// }
 
-  // Card.findByIdAndRemove(req.params.cardId)
-  //   .then((card) => ((!card) ? responseNotFound(res, 'Карточка не найдена') : res.send(card)))
-  //   .catch((err) => {
-  //     if (err.name === 'ValidationError' || err.name === 'CastError') {
-  //       responseBadRequestError(res);
-  //     } else {
-  //       responseServerError(res);
-  //     }
-  //   });// данные не записались, вернём ошибку
+export function deleteCard(req, res, next) {
+  Card.findById(req.params.id)
+    .then((card) => {
+      if (!card) {
+        responseNotFound(res, 'Карточка не найдена');
+      } else if (card.owner !== req.user._id) {
+        throw forbiddenError;
+      } else {
+        Card.findOneAndRemove({ _id: req.params.id, owner: req.user._id })
+          .then((card) => ((!card) ? responseNotFound(res, 'Карточка не найдена, или это не карточка другого пользователя') : res.send(card)))
+          .catch((err) => {
+            if (err.name === 'ValidationError' || err.name === 'CastError') {
+              responseBadRequestError(res);
+            } else {
+              responseServerError(res);
+            }
+        });// данные не записались, вернём ошибку
+      }
+    })
+    .catch((err) => {
+      if (err instanceof HTTPError) {
+        next(err);
+      } else if (err.name === 'CastError') {
+        next(buildErrorBadRequest(err.message));
+      } else {
+        next(buildErrorServer(err.message));
+      }
+    });
 }
+
+// export function deleteCard(req, res) {
+//   Card.findOneAndRemove({ _id: req.params.id, owner: req.user._id })
+//     .then((card) => ((!card) ? responseNotFound(res, 'Карточка не найдена, или это не карточка другого пользователя') : res.send(card)))
+//     .catch((err) => {
+//       if (err.name === 'ValidationError' || err.name === 'CastError') {
+//         responseBadRequestError(res);
+//       } else {
+//         responseServerError(res);
+//       }
+//     });// данные не записались, вернём ошибку
+// }
 
 export function likeCard(req, res) {
   Card.findByIdAndUpdate(
